WIP: Refactor cronjob-network code to cope with unexpected null-jobs
Null-jobs occur when the backend removes a throw-away cron-job (e.g. for scenes) but there is an unexpected behaviour
in Gun that those not always get filtered out in map().
This leaves us unable to remove the job from our local store since all we get is a null-tombstone without a job-id.

The temporary workaround is now to emit an error and skip the cronjob-processing logic when we receive a null-cronjob.